refactor: share the pass-through attribute declarations with their forwarding
The new MathBaseListOperator copied the block of `leaveRaw` declarations for
`fixed` and the rounding attributes out of Sequence, leaving the list of
forwarded attributes stated in three places: twice as declarations and once in
`returnPassThroughAttributes`. Declare them from a single helper next to the
forwarding, so a composite cannot declare one set and forward another.

Also correct two pieces of prose: the state variables ListIndexBaseOperator
drops match `<periodicSet>`, not `<evaluate>`/`<matrix>` (both of which keep an
`unordered` state variable); and document that `<searchSorted>` with no target
gives 0, which is otherwise not a position it can return.
