+++ packages/doenetml-worker-javascript/src/utils/valueListReplacements.js
@@ -27,13 +27,38 @@ import { returnNumberDisplayAttributes } from "./numberDisplay";
  * results. `fixed` is included so an author can override the `fixed="true"`
  * that these replacements otherwise carry — their values are computed, so they
  * are not modifiable by default.
+ *
+ * The composite must declare them all as `leaveRaw`, since it forwards them
+ * rather than acting on them itself. Since the declarations and the forwarding
+ * have to name the same attributes, both come from here.
+ */
+export function returnPassThroughAttributeDeclarations() {
+    let attributes = {
+        fixed: {
+            leaveRaw: true,
+            description:
+                "Whether this component's value is fixed and cannot be modified.",
+        },
+    };
+
+    const numberDisplayAttrs = returnNumberDisplayAttributes();
+    for (let attrName in numberDisplayAttrs) {
+        attributes[attrName] = {
+            leaveRaw: true,
+            description: numberDisplayAttrs[attrName].description,
+        };
+    }
+
+    return attributes;
+}
+
+/**
+ * The pass-through attributes an author actually wrote on `component`, ready to
+ * be converted onto its replacements.
  */
 export function returnPassThroughAttributes(component) {
     let attributesToConvert = {};
-    for (let attr of [
-        "fixed",
-        ...Object.keys(returnNumberDisplayAttributes()),
-    ]) {
+    for (let attr of Object.keys(returnPassThroughAttributeDeclarations())) {
         if (attr in component.attributes) {
             attributesToConvert[attr] = component.attributes[attr];
         }
